% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_reticulate.R
\name{adata_to_seurat}
\alias{adata_to_seurat}
\title{Convert adata to a Seurat object}
\usage{
adata_to_seurat(adata, counts_layer, ...)
}
\arguments{
\item{adata}{An `anndata` python object}

\item{counts_layer}{layer of adata containing counts information}
}
\value{
A converted Seurat object
}
\description{
This function reads a h5ad file, extracts metadata, gene and cell names, and count data.
Then, it creates a Seurat object with the extracted information.
}
\details{
Note that this version doesn't import unstructured data saved in `uns`.
}
