% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogenes.R
\name{autogenes_deconvolute}
\alias{autogenes_deconvolute}
\title{AutoGeneS deconvolution of bulk data using an `screference`}
\usage{
autogenes_deconvolute(
  bulk_data,
  scref,
  model = c("nusvr", "nnls", "linear")[1]
)
}
\arguments{
\item{bulk_data}{a matrix of genes-by-samples with bulk mixtures}

\item{scref}{an object of class `screference`}

\item{model}{a model to run AutoGeneS. One of: `'nusvr'`, `'nnls'` or `'linear'`}
}
\value{
a tibble with deconvolution fractions
}
\description{
AutoGeneS deconvolution of bulk data using an `screference`
}
\note{
Reference: Aliee, Hananeh, and Fabian J. Theis. 2021.
“AutoGeneS: Automatic Gene Selection Using Multi-Objective Optimization
for RNA-Seq Deconvolution.” Cell Systems 12 (7): 706-715.e4.
https://doi.org/10.1016/j.cels.2021.05.006.

See also: https://github.com/theislab/AutoGeneS
}
