% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogenes.R
\name{autogenes_scref}
\alias{autogenes_scref}
\title{Compute reference matrix from an `screference` object using
AutoGeneS}
\usage{
autogenes_scref(
  scref,
  ngen = 500,
  mode = "standard",
  nfeatures = 2000,
  seed = 0,
  population_size = 100,
  offspring_size = 50,
  crossover_pb = 0.7,
  mutation_pb = 0.3,
  crossover_thresh = 1000,
  int_standard_pb = 0.1,
  ...
)
}
\arguments{
\item{scref}{an object of class `screference`}

\item{ngen}{an integer, number of generations for optimization}

\item{mode}{one of `"standard"` or `"fixed"`. If "fixed", `"nfeatures"` is used.}

\item{nfeatures}{integer, number of genes selected in fixed mode.}

\item{seed}{int, for reproducibility}

\item{population_size}{int, size of each generation (mu parameter)}

\item{offspring_size}{int, number of individuals per generation (lambda parameter)}

\item{crossover_pb}{float, crossover probability}

\item{mutation_pb}{float, mutation probability}

\item{...}{passed to [Seurat::FindVariableFeatures]}

\item{crossover_thres}{int, crossover threshold (standard mode)}

\item{ind_standard_pb}{float, probability used to generate initial population in
standard mode}
}
\value{
an screference object updated with AutoGeneS reference centroids
}
\description{
Compute reference matrix from an `screference` object using
AutoGeneS
}
\note{
Reference: Aliee, Hananeh, and Fabian J. Theis. 2021.
“AutoGeneS: Automatic Gene Selection Using Multi-Objective Optimization
for RNA-Seq Deconvolution.” Cell Systems 12 (7): 706-715.e4.
https://doi.org/10.1016/j.cels.2021.05.006.

See also: https://github.com/theislab/AutoGeneS
}
