% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesprism.R
\name{bayesprism_deconvolute}
\alias{bayesprism_deconvolute}
\title{BayesPrism deconvolution of bulk data using an `screference`}
\usage{
bayesprism_deconvolute(
  bulk_data,
  scref,
  cache_path = NULL,
  outlier_cut = 0.01,
  outlier_fraction = 0.1,
  pseudo_min = 1e-08,
  ncores = parallel::detectCores()/2
)
}
\arguments{
\item{bulk_data}{a matrix of genes-by-samples with bulk mixtures}

\item{scref}{an object of class `screference`}

\item{cache_path}{path to cache the results}

\item{outlier_cut, outlier.fraction}{two floats used to filter genes in `bulk_data`
whose expression fraction is greater than outlier.cut in more than outlier.fraction.
Typically for dataset with reasonable quality control, very few genes will be
filtered. Removal of outlier genes will ensure that the inference will not
be dominated by outliers, which sometimes may be resulted from poor QC in mapping.
See: [BayesPrism::new.prism()]}

\item{pseudo_min}{float, the desired minimum value to replace zero
after normalization. See: [BayesPrism::new.prism()].}

\item{n_cores}{number of cores used for computation}
}
\value{
a tibble with deconvolution fractions
}
\description{
BayesPrism deconvolution of bulk data using an `screference`
}
\note{
Reference: Chu, T., Wang, Z., Pe’er, D. et al. Cell type and gene
expression deconvolution with BayesPrism enables Bayesian integrative
analysis across bulk and single-cell RNA sequencing in oncology.
Nat Cancer 3, 505–517 (2022). https://doi-org.insb.bib.cnrs.fr/10.1038/s43018-022-00356-3
See also: https://github.com/Danko-Lab/BayesPrism
}
