% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesprism.R
\name{bayesprism_scref}
\alias{bayesprism_scref}
\title{Compute reference matrix from an `screference` object using
BayesPrism}
\usage{
bayesprism_scref(
  scref,
  cache_path = "bayes_prism",
  data_type = c("10X", "Smart-seq")[1],
  malignant_pop_id = NULL,
  pval_cutoff = 0.01,
  logFC_cutoff = 0.1,
  ncores = parallel::detectCores()/2
)
}
\arguments{
\item{scref}{an object of class `screference`}

\item{cache_path}{path to cache the results}

\item{data_type}{a string, one of: `"10X"` or `"Smart-seq"`}

\item{malignant_pop_id}{the ID in the screference `annot_id` corresponding
to the malignant cell population for deconvolution of cancer samples. If none
of the populations are malignant, leave as `NULL`.}

\item{pval_cutoff}{a float, p-value cutoff from differential expression
analysis for selection of markers}

\item{logFC_cutoff}{a float, cutoff for log fold-change from differential
expression analysis between populations for selection of markers}

\item{n_cores}{number of cores used for computation}
}
\value{
an screference object updated with BayesPrism reference
}
\description{
Compute reference matrix from an `screference` object using
BayesPrism
}
\note{
Reference: Chu, T., Wang, Z., Pe’er, D. et al. Cell type and gene
expression deconvolution with BayesPrism enables Bayesian integrative
analysis across bulk and single-cell RNA sequencing in oncology.
Nat Cancer 3, 505–517 (2022). https://doi-org.insb.bib.cnrs.fr/10.1038/s43018-022-00356-3
See also: https://github.com/Danko-Lab/BayesPrism
}
