% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisque.R
\name{bisque_deconvolute}
\alias{bisque_deconvolute}
\title{Bisque deconvolution of bulk data using an `screference`}
\usage{
bisque_deconvolute(bulk_data, scref)
}
\arguments{
\item{bulk_data}{a matrix of bulk RNA-seq data with rows representing genes
and columns representing samples.}

\item{scref}{an object of class `screference`}
}
\value{
a tibble containing the fraction of each cell type in each bulk sample.
}
\description{
Bisque deconvolution of bulk data using an `screference`
}
\note{
Reference:
Jew, B., Alvarez, M., Rahmani, E. et al. Accurate estimation of cell
composition in bulk expression through robust integration of single-cell
information. Nat Commun 11, 1971 (2020). https://doi.org/10.1038/s41467-020-15816-6
See also: https://github.com/cozygene/bisque
}
