% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card_deconvolute}
\alias{card_deconvolute}
\title{CARD deconvolution using an `screference`}
\usage{
card_deconvolute(spatial_obj, scref)
}
\arguments{
\item{spatial_obj}{an `Seurat` object with a Spatial assay}

\item{scref}{an `screference` object containing single-cell reference data
and pre-computed references}
}
\description{
CARD deconvolution using an `screference`
}
