% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card_scref}
\alias{card_scref}
\title{Construct the mean gene expression basis matrix (B) from `screference`
object using the method implemented in the CARD package}
\usage{
card_scref(scref)
}
\arguments{
\item{scref}{an object of `screference`}
}
\value{
a Basis matrix
}
\description{
Construct the mean gene expression basis matrix (B) from `screference`
object using the method implemented in the CARD package
}
