% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cibersortx.R
\name{cibersortx_deconvolute}
\alias{cibersortx_deconvolute}
\title{CIBERSORTx deconvolution of bulk data using an `screference`}
\usage{
cibersortx_deconvolute(
  bulk_data,
  scref,
  cache_path = "cibersortx",
  username = NULL,
  token = NULL
)
}
\arguments{
\item{bulk_data}{a matrix of genes-by-samples with bulk mixtures}

\item{scref}{an object of class `screference`}

\item{cache_path}{path to cache deconvolution results}

\item{username}{a username used to authenticate access to CIBERSORTx method.
See: https://cibersortx.stanford.edu/getoken.php}

\item{token}{a token used to authenticate access to CIBERSORTx method}

\item{a}{tibble with deconvolution fractions}
}
\description{
CIBERSORTx deconvolution of bulk data using an `screference`
}
