% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cibersortx.R
\name{cibersortx_scref}
\alias{cibersortx_scref}
\title{Compute reference signature matrix from `screference` object using
method implemented in CIBERSORTx}
\usage{
cibersortx_scref(
  scref,
  cache_path = "cibersortx",
  username = NULL,
  token = NULL
)
}
\arguments{
\item{scref}{an object of `screference}

\item{cache_path}{path to cache results}

\item{username}{a username used to authenticate access to CIBERSORTx method.
See: https://cibersortx.stanford.edu/getoken.php}

\item{token}{a token used to authenticate access to CIBERSORTx method}
}
\value{
a path to cached signature matrix
}
\description{
Compute reference signature matrix from `screference` object using
method implemented in CIBERSORTx
}
