% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screference.R
\name{compute_reference}
\alias{compute_reference}
\title{Compute reference given a deconvolution method from an `screference` object}
\usage{
compute_reference(scref, method = c("cibersortx", "dwls")[1], ...)
}
\arguments{
\item{scref}{an `screference` object}

\item{method}{a string, one of `deconvolution_methods()`. Only CIBERSORTx
and methods implemented in the DWLS package (`dwls`, `svr`, `ols`) need
to pre-compute reference.}

\item{...}{passed to deconvolution method's reference computation function.
See the wrapper (`{method}_scref`) where `{method}` is the method name for
parameters.}
}
\value{
an object of class `screference`
}
\description{
Compute reference given a deconvolution method from an `screference` object
}
