% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconverse-package.R
\docType{package}
\name{deconverse-package}
\alias{deconverse}
\alias{deconverse-package}
\title{deconverse: Benchmarking of bulk deconvolution methods}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
