% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R, R/scbench.R
\name{deconverse_results.Seurat}
\alias{deconverse_results.Seurat}
\alias{deconverse_results.scbench}
\title{Returns feature names of `deconverse` results that are stored in the Seurat object}
\usage{
\method{deconverse_results}{Seurat}(
  spatial_obj,
  method = NULL,
  major_population = FALSE,
  raw_results = FALSE
)

\method{deconverse_results}{scbench}(scbench, tidy = FALSE, ...)
}
\arguments{
\item{spatial_obj}{a Seurat object with a "Spatial" assay}

\item{method}{a string, corresponding to the method of deconvolution.
Must be one of `deconvolution_methods()`. If NULL, all available
deconvolution features are returned.}

\item{major_population}{if TRUE, the method returns a categorical
feature with the most abundant population predicted for each spot.}

\item{raw_results}{if TRUE, returns deconverse results as stored in the object
instead of feature names. Also available on `spatial_obj@tools$deconverse`}
}
\value{
a vector of strings correspoding to features in the Seurat object.
}
\description{
Returns feature names of `deconverse` results that are stored in the Seurat object
}
