% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{deconvolute.matrix}
\alias{deconvolute.matrix}
\title{Deconvolute bulk RNA-seq matrix using a chosen method}
\usage{
deconvolute.matrix(
  bulk_data,
  scref,
  method = deconvolution_methods()[1],
  bulk_norm = c("rpm", "none", "proportional_fitting")[3],
  ...
)
}
\arguments{
\item{bulk_data}{a count or linear-normalized (cpm, etc) matrix of genes-by-samples}

\item{scref}{an `screference` object containing single-cell reference data
and pre-computed references for methods that require them.}

\item{method}{a string with the name of the deconvolution method. For available
methods, consult `deconvolution_methods()`}

\item{bulk_norm}{the normalization to be applied in the bulk data. One of:
`"rpm"` for reads per million, `"none"` for raw counts, or `"proportional_fitting"`
for using the mean library size of the pseudobulk data to normalize. If already
normalized, make sure to choose `"none"`.}

\item{...}{other parameters, passed to the method wrapper, enables the user
to change the method parameters. See: `deconvolute_{method}` where `{method}`
is the method name in lowercase for method-specific parameters}
}
\value{
an object of class `scbench`
}
\description{
Deconvolute bulk RNA-seq matrix using a chosen method
}
