% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{deconvolute_all}
\alias{deconvolute_all}
\title{Deconvolute `scbench` object using all methods with default settings}
\usage{
deconvolute_all(scbench, scref, methods = deconvolution_methods(), ...)
}
\arguments{
\item{scbench}{an `scbench` object already processed by `pseudobulks`}

\item{scref}{an `screference` object containing references for all methods that
require them}

\item{methods}{a vector of strings, which methods to use. For available
methods, consult `deconvolution_methods()`}

\item{...}{passed to `deconvolute` for CIBERSORTx credentials}
}
\value{
an object of class `scbench`
}
\description{
Deconvolute `scbench` object using all methods with default settings
}
