% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screference.R
\name{delete_cached}
\alias{delete_cached}
\title{Delete cached data on an `screference` object}
\usage{
delete_cached(scref, which)
}
\arguments{
\item{scref}{an `screference` object}

\item{which}{a string, which cached result to delete}
}
\value{
screference object
}
\description{
Delete cached data on an `screference` object
}
