% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwls.R
\name{dwls_deconvolute}
\alias{dwls_deconvolute}
\title{Dampened weighted least squares (DWLS) deconvolution of bulk data using an `screference`}
\usage{
dwls_deconvolute(bulk_data, scref, ncores = 4)
}
\arguments{
\item{bulk_data}{a matrix of genes-by-samples with bulk mixtures}

\item{scref}{an object of class `screference`}

\item{n_cores}{number of cores used for computation}
}
\value{
a tibble with deconvolution fractions
}
\description{
Dampened weighted least squares (DWLS) deconvolution of bulk data using an `screference`
}
