% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwls.R
\name{dwls_scref}
\alias{dwls_scref}
\title{Compute reference signature matrix from `screference` object using
method implemented in DWLS package for methods `dwls`, `svr` and `ols`}
\usage{
dwls_scref(scref, cache_path = "dwls", logFC_cutoff = 0.5, pval_cutoff = 0.01)
}
\arguments{
\item{scref}{an object of `screference}

\item{cache_path}{path to cache results}

\item{logFC_cutoff}{a float, cutoff for log fold-change from differential
expression analysis between populations for selection of markers}

\item{pval_cutoff}{a float, p-value cutoff from differential expression
analysis for selection of markers}
}
\value{
a signature matrix
}
\description{
Compute reference signature matrix from `screference` object using
method implemented in DWLS package for methods `dwls`, `svr` and `ols`
}
