% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{get_benchmark_results}
\alias{get_benchmark_results}
\title{Get benchmark results from `scbench` object}
\usage{
get_benchmark_results(
  scbench,
  methods = NULL,
  level = NULL,
  type = "population"
)
}
\arguments{
\item{scbench}{an `scbench` object already processed by `deconvolute` or
`deconvolute_all`}

\item{methods}{NULL, a string, or vector of strings. Results will be returned
for the methods indicated. See `deconvolution_methods()` for methods.}

\item{type}{type of deconvoluted mixtures. One of: `"population"`,
`"spillover"` or `"lod"`}

\item{a}{`tibble` with results.}
}
\description{
Get benchmark results from `scbench` object
}
