% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cibersortx.R
\name{install_cibersortx}
\alias{install_cibersortx}
\title{Installs CIBERSORTx container using docker or singularity}
\usage{
install_cibersortx(container_type = c("docker", "singularity")[1])
}
\arguments{
\item{container_type}{container software used. one of:
"docker" or "singularity"}
}
\description{
Installs CIBERSORTx container using docker or singularity
}
