% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{lod_proportions}
\alias{lod_proportions}
\title{Simulate proportions to find the limit of detection of each population in
mixtures in an `scbench` object}
\usage{
lod_proportions(scbench, max_prop = 0.1, step = 0.005)
}
\arguments{
\item{scbench}{an `scbench` object}

\item{max_prop}{a float, the maximum proportion of each target population
tested}

\item{step}{the step between 0 and `max_prop` in a sequence to be tested as
limit of detection}
}
\value{
an object of class `scbench`
}
\description{
Simulate proportions to find the limit of detection of each population in
mixtures in an `scbench` object
}
