% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{mixtures_population}
\alias{mixtures_population}
\title{Simulate sample population mixtures given bounds in an `scbench` object}
\usage{
mixtures_population(scbench, nsamps = 1000, seed = 0)
}
\arguments{
\item{scbench}{an `scbench` object}

\item{nsamps}{an integer representing the number of samples to simulate}

\item{seed}{an integer representing a "seed", for reproducibility of mixtures.}
}
\value{
an object of class `scbench`
}
\description{
Simulate sample population mixtures given bounds in an `scbench` object
}
