% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{mixtures_spatial}
\alias{mixtures_spatial}
\title{Simulate spatially-aware mixtures in an `scbench` object}
\usage{
mixtures_spatial(
  scbench,
  n_samples = 20,
  n_spots = 2048,
  n_zones = 4,
  majority_var = 0.15,
  min_majority_prop = 0.3,
  seed = 1,
  ncores = 8,
  ...
)
}
\arguments{
\item{scbench}{an `scbench` object}

\item{n_samples}{an integer representing the number of samples to simulate}

\item{n_spots}{an integer representing the number of spots to simulate within each sample}

\item{n_zones}{the number of different zones to generate per slide with different
transcriptomic profiles}

\item{majority_var}{variation parameter related to how much the "majority" population
proportion varies within a zone (range: 0 - 0.3)}

\item{min_majority_prop}{minimum proportion of the "majority" population in a
zone}

\item{seed}{an integer representing a "seed", for reproducibility of mixtures.}

\item{...}{arguments passed to `fields::circulantEmbeddingSetup` related to
the simulation of the Random Gaussian Field}
}
\value{
an object of class `scbench`
}
\description{
Simulate spatially-aware mixtures in an `scbench` object
}
