% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{mixtures_spillover}
\alias{mixtures_spillover}
\title{Simulate mixtures for measuring spillover between all pairs of populations
in an `scbench` object}
\usage{
mixtures_spillover(scbench, step = 0.05)
}
\arguments{
\item{scbench}{an `scbench` object}

\item{step}{a float, defines the step in the sequence for creating mixtures
and simulating spillover effects}
}
\value{
an object of class `scbench`
}
\description{
Simulate mixtures for measuring spillover between all pairs of populations
in an `scbench` object
}
