% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_scdc.R
\name{music_deconvolute}
\alias{music_deconvolute}
\title{MuSiC deconvolution of bulk data using an `screference`}
\usage{
music_deconvolute(bulk_data, scref)
}
\arguments{
\item{bulk_data}{a matrix of genes-by-samples with bulk mixtures}

\item{scref}{an object of class `screference`}

\item{a}{tibble with deconvolution fractions and model metrics}
}
\description{
MuSiC deconvolution of bulk data using an `screference`
}
