% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screference.R
\name{new_hscreference}
\alias{new_hscreference}
\title{Generate multi-level hierarchical single-cell reference (`hscreference`) object}
\usage{
new_hscreference(
  seurat_obj,
  annot_ids,
  batch_id = NULL,
  project_name = "project",
  cache_path = "scref_cache",
  sample_cells = NULL,
  seed = NULL
)
}
\arguments{
\item{seurat_obj}{a reference Seurat object}

\item{annot_ids}{in order, column names of the Seurat object metadata from
coarsest to finest grained annotation}

\item{batch_id}{NULL or a column name of the Seurat object metadata containing
information about batch to be used as batch information downstream. Examples
of batch are patient ID, study ID, etc.}

\item{project_name}{a string indicating the project name, used for caching}

\item{cache_path}{path to the directory where results will be cached}

\item{sample_cells}{NULL or a list with how many cells to randomly sub-sample of each
population in the coarsest grained annotation.
Sub-sampling can make the object significantly smaller and accelerate
downstream computation, but make the results less accurate.}

\item{seed}{an integer representing a "seed", for reproducibility of sub-sampling.}
}
\value{
an object of class `hscreference`
}
\description{
Generate multi-level hierarchical single-cell reference (`hscreference`) object
}
