% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screference.R
\name{new_screference}
\alias{new_screference}
\title{Generate `screference` object}
\usage{
new_screference(
  seurat_obj,
  annot_id,
  batch_id = NULL,
  project_name = "project",
  cache_path = "scref_cache",
  sample_cells = NULL,
  seed = NULL
)
}
\arguments{
\item{seurat_obj}{a reference Seurat object}

\item{annot_id}{the column name of the Seurat object metadata containing
the reference annotation}

\item{batch_id}{NULL or a column name of the Seurat object metadata containing
information about batch to be used as batch information downstream. Examples
of batch are patient ID, study ID, etc.}

\item{project_name}{a string indicating the project name, used for caching}

\item{cache_path}{path to the directory where results will be cached}

\item{sample_cells}{NULL or a list with how many cells to randomly sub-sample of each
population. Sub-sampling can make the object significantly smaller and accelerate
downstream computation, but make the results less accurate.}

\item{seed}{an integer representing a "seed", for reproducibility of sub-sampling.}
}
\value{
an object of class `screference`
}
\description{
Generate `screference` object
}
