% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwls.R
\name{ols_deconvolute}
\alias{ols_deconvolute}
\title{Ordinary least squares (OLS) deconvolution of bulk data using an `screference`}
\usage{
ols_deconvolute(bulk_data, scref, ncores = 4)
}
\arguments{
\item{bulk_data}{a matrix of genes-by-samples with bulk mixtures}

\item{scref}{an object of class `screference`}

\item{n_cores}{number of cores used for computation}

\item{a}{tibble with deconvolution fractions}
}
\description{
Ordinary least squares (OLS) deconvolution of bulk data using an `screference`
}
