% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc}
\alias{pbmc}
\title{Seurat example PBCM dataset}
\format{
`pbmc`
A Seurat Object.
}
\source{
<https://satijalab.org/seurat/articles/pbmc3k_tutorial.html>
}
\usage{
pbmc
}
\description{
Cells annotated with the example pipeline from the Seurat package vignette,
already preprocessed, and with Platelet cluster removed (too few cells)
}
\keyword{datasets}
