% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_plots.R
\name{plt_colocalization}
\alias{plt_colocalization}
\title{Plot spatial correlation of cell type proportions}
\usage{
plt_colocalization(spatial_obj, method = NULL, level = NULL, min_prop = 0.01)
}
\arguments{
\item{spatial_obj}{an `Seurat` object with a Spatial assay with `deconverse::deconvolute` results}

\item{method}{a method to plot. If NULL and only one method ran, that method will be plotted.}

\item{level}{a level of annotation to plot. If NULL, the finest level will be plotted.}

\item{min_prop}{minimal proportion of a cell type to be considered as occuring in a given spot}
}
\value{
a ggnetwork plot of co-occurance of cell types in spots
}
\description{
Plot spatial correlation of cell type proportions
}
