% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R, R/screference.R
\name{plt_comp_performance.scbench}
\alias{plt_comp_performance.scbench}
\alias{plt_comp_performance.hscreference}
\alias{plt_comp_performance.screference}
\title{Plot computational metrics computed during deconvolution}
\usage{
\method{plt_comp_performance}{scbench}(scbench, types = NULL, levels = NULL)

\method{plt_comp_performance}{hscreference}(hscref)

\method{plt_comp_performance}{screference}(scref)
}
\arguments{
\item{scbench}{an `scbench` object that has been evaluated by
`deconvolute`}

\item{types}{type of mixtures to deconvolute. One of: `"population"`,
`"spillover"` or `"lod"`. If NULL (default), all types are plotted.}

\item{levels}{annotation levels to show. If NULL (default), all levels are plotted.}

\item{hscref}{an `h-screference` object}

\item{scref}{an `screference` object}
}
\value{
a patchwork plot of time elapsed and peak RAM consumption.

a patchwork plot of time elapsed and peak RAM consumption.

a patchwork plot of time elapsed and peak RAM consumption.
}
\description{
Plot computational metrics computed during deconvolution

Plot computational metrics computed during single-cell reference computation

Plot computational metrics computed during single-cell reference computation
}
