% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{plt_cors_scatter}
\alias{plt_cors_scatter}
\title{Plot correlations by population between deconvolution results
for simulated populations and true proportions used for the pseudobulk}
\usage{
plt_cors_scatter(scbench, method)
}
\arguments{
\item{scbench}{an `scbench` object that has been evaluated by
`deconvolute`}

\item{method}{a string, one of `deconvolution_methods()`}
}
\value{
a ggplot object
}
\description{
Plot correlations by population between deconvolution results
for simulated populations and true proportions used for the pseudobulk
}
