% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{plt_lod_heatmap}
\alias{plt_lod_heatmap}
\title{Plot heatmap summarizing limit of detection by population and method}
\usage{
plt_lod_heatmap(scbench)
}
\arguments{
\item{scbench}{an `scbench` object that has been evaluated by
`deconvolute`}
}
\value{
a list, containing `heatmap`, a ggheatmap object, and `lod_table`, a
table summarizing results
}
\description{
Plot heatmap summarizing limit of detection by population and method
}
