% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{plt_lod_scatter}
\alias{plt_lod_scatter}
\title{Plot limit of detection of deconvolution for each population as the proportion
of the pseudobulk where the method estimation is higher than at proportion=0 for
each population}
\usage{
plt_lod_scatter(scbench, method)
}
\arguments{
\item{scbench}{an `scbench` object that has been evaluated by
`deconvolute`}

\item{method}{a string, one of `deconvolution_methods()`}
}
\value{
a ggplot object
}
\description{
Plot limit of detection of deconvolution for each population as the proportion
of the pseudobulk where the method estimation is higher than at proportion=0 for
each population
}
