% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_plots.R
\name{plt_method_concordance}
\alias{plt_method_concordance}
\title{Plot percentage of concordance of major population between methods on spatial data}
\usage{
plt_method_concordance(spatial_obj)
}
\arguments{
\item{spatial_obj}{an `Seurat` object with a Spatial assay with `deconverse::deconvolute` results}
}
\value{
a list containing the plot and table for the comparisons
}
\description{
Plot percentage of concordance of major population between methods on spatial data
}
