% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_plots.R
\name{plt_method_correlation}
\alias{plt_method_correlation}
\title{Plot correlation between methods on spatial data}
\usage{
plt_method_correlation(
  spatial_obj,
  level = 1,
  nrow = 1,
  cor_fill_limits = c(0, 1),
  cor_text_size = 3
)
}
\arguments{
\item{spatial_obj}{an `Seurat` object with a Spatial assay with `deconverse::deconvolute` results}

\item{level}{a level of annotation to plot. If NULL, the finest level will be plotted.}

\item{nrow}{number of rows of the plot facets}

\item{cor_fill_limits}{a vector with two numbers, representing the lower and
upper limits of the color scale for the correlation}

\item{cor_text_size}{a number, representing the size of the text of the
correlation in the plot tiles.}
}
\value{
a list containing the plot and table for the comparisons
}
\description{
Plot correlation between methods on spatial data
}
