% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{plt_population_proportions}
\alias{plt_population_proportions}
\title{Plot sample population proportions from an `scbench` object}
\usage{
plt_population_proportions(scbench, nshow = 50, order_by = NULL)
}
\arguments{
\item{scbench}{an `scbench` object that has been evaluated by
`population_proportions`}

\item{nshow}{an integer, the number of samples to plot}

\item{order_by}{a string of a name of a population to order the samples with
increasing proportions of `order_by`}
}
\value{
a ggplot object
}
\description{
Plot sample population proportions from an `scbench` object
}
