% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_plots.R
\name{plt_spatial_correlation}
\alias{plt_spatial_correlation}
\title{Plot spatial correlation of cell type proportions}
\usage{
plt_spatial_correlation(spatial_obj, method = NULL, level = NULL)
}
\arguments{
\item{spatial_obj}{an `Seurat` object with a Spatial assay with `deconverse::deconvolute` results}

\item{method}{a method to plot. If NULL and only one method ran, that method will be plotted.}

\item{level}{a level of annotation to plot. If NULL, the finest level will be plotted.}
}
\value{
a ggplot of spatial correlation between cell types
}
\description{
Plot spatial correlation of cell type proportions
}
