% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{plt_spatial_mixture}
\alias{plt_spatial_mixture}
\title{Plot sample spatial mixture from an `scbench` object}
\usage{
plt_spatial_mixture(scbench, nslide = 1, level = NULL)
}
\arguments{
\item{scbench}{an `scbench` object that has been evaluated by
`mixtures_spatial`}

\item{nslide}{an integer, the number of the slide to plot}

\item{level}{what level of annotation to plot. If not specified,
finest will be shown.}
}
\value{
a ggplot object
}
\description{
Plot sample spatial mixture from an `scbench` object
}
