% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{plt_spillover_heatmap}
\alias{plt_spillover_heatmap}
\title{Plot heatmap summarizing spillover RMSE by population and method}
\usage{
plt_spillover_heatmap(scbench, level = NULL)
}
\arguments{
\item{scbench}{an `scbench` object that has been evaluated by
`deconvolute`}

\item{level}{an annotation level to show. Finest-grained annotation (highest level)
by default.}
}
\value{
a list, containing `heatmap`, a ggheatmap object, and `rmse_table`, a
table summarizing results
}
\description{
Plot heatmap summarizing spillover RMSE by population and method
}
