% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{plt_spillover_scatter}
\alias{plt_spillover_scatter}
\title{Plot correlations between deconvolution results and spillover proportions
between each pair of populations}
\usage{
plt_spillover_scatter(scbench, method)
}
\arguments{
\item{scbench}{an `scbench` object that has been evaluated by
`deconvolute`}

\item{method}{a string, one of `deconvolution_methods()`}
}
\value{
a ggplot object
}
\description{
Plot correlations between deconvolution results and spillover proportions
between each pair of populations
}
