% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{population_proportions}
\alias{population_proportions}
\title{Simulate sample population proportions given bounds in an `scbench` object}
\usage{
population_proportions(scbench, nsamps = 1000, seed = NULL)
}
\arguments{
\item{scbench}{an `scbench` object}

\item{nsamps}{an integer representing the number of samples to simulate}

\item{seed}{an integer representing a "seed", for reproducibility of proportions.}
}
\value{
an object of class `scbench`
}
\description{
Simulate sample population proportions given bounds in an `scbench` object
}
