% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{pseudobulks}
\alias{pseudobulks}
\title{Generate pseudobulk gene expression from single-cell RNA-seq given
population mixtures in an `scbench` object}
\usage{
pseudobulks(
  scbench,
  ncells = 2000,
  ncores = 8,
  shrink_obj = TRUE,
  by_batch = FALSE,
  seed = 0
)
}
\arguments{
\item{scbench}{an `scbench` object that has been evaluated by
`population_mixtures`}

\item{ncells}{an integer, the number of cells to pool into pseudobulk.}

\item{shrink_obj}{a boolean, if TRUE, `scbench` object is reduced in size by
removing the single-cell expression data and keeping only its metadata after
estimating the pseudobulks}

\item{by_batch}{a boolean, if TRUE, pseudobulks are estimated by batch using
the `batch_id` provided when creating the object.}

\item{seed}{an integer representing a "seed", for reproducibility of mixtures.}

\item{level}{a string character, the reference data level for pseudobulk pooling
(e.g. `"l1"`, `"l2"`)}
}
\value{
an object of class `scbench`
}
\description{
Generate pseudobulk gene expression from single-cell RNA-seq given
population mixtures in an `scbench` object
}
