% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctd.R
\name{rctd_deconvolute}
\alias{rctd_deconvolute}
\title{RCTD deconvolution using an `screference`}
\usage{
rctd_deconvolute(
  spatial_obj,
  scref,
  ncores = 4,
  gene_cutoff = 0.000125,
  fc_cutoff = 0.5,
  gene_cutoff_reg = 2e-04,
  fc_cutoff_reg = 0.75,
  UMI_min = 100,
  UMI_max = 2e+07,
  counts_MIN = 10,
  UMI_min_sigma = 300,
  CELL_MIN_INSTANCE = 25,
  MAX_MULTI_TYPES = 4,
  CONFIDENCE_THRESHOLD = 5,
  DOUBLET_THRESHOLD = 20,
  cache_path = "rctd"
)
}
\arguments{
\item{spatial_obj}{an `Seurat` object with a Spatial assay}

\item{scref}{an object of `screference`}

\item{gene_cutoff}{minimum normalized gene expression for genes to be included in the platform effect normalization step.}

\item{fc_cutoff}{minimum log-fold-change (across cell types) for genes to be included in the platform effect normalization step.}

\item{gene_cutoff_reg}{minimum normalized gene expression for genes to be included in the RCTD step.}

\item{fc_cutoff_reg}{minimum log-fold-change (across cell types) for genes to be included in the RCTD step.}

\item{UMI_min}{minimum UMI per pixel included in the analysis}

\item{UMI_max}{maximum UMI per pixel included in the analysis}

\item{counts_MIN}{minimum total counts per pixel of genes used in the analysis.}

\item{UMI_min_sigma}{minimum UMI per pixel for the \link{choose_sigma_c} function}

\item{CELL_MIN_INSTANCE}{minimum number of cells required per cell type. Default 25, can be lowered if desired.}

\item{MAX_MULTI_TYPES}{max number of cell types per pixel}

\item{CONFIDENCE_THRESHOLD}{the minimum change in likelihood (compared to other cell types) necessary to determine a cell type identity with confidence}

\item{DOUBLET_THRESHOLD}{the penalty weight of predicting a doublet instead of a singlet for a pixel}

\item{cache_path}{a path to cache the results}

\item{max_cores}{for parallel processing, the number of cores used. If set to 1, parallel processing is not used. The system will additionally be checked for
number of available cores.}
}
\value{
a tibble with deconvolution fractions
}
\description{
RCTD deconvolution using an `screference`
}
