% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_reticulate.R
\name{read_h5ad}
\alias{read_h5ad}
\title{Read a h5ad file and convert it to a Seurat object}
\usage{
read_h5ad(file, counts_layer = "counts", ...)
}
\arguments{
\item{file}{A character string representing the path to the h5ad file.}
}
\value{
A Seurat object containing the data from the h5ad file.
}
\description{
This function reads a h5ad file, extracts metadata, gene and cell names, and count data.
Then, it creates a Seurat object with the extracted information.
}
\details{
Note that this version doesn't import unstructured data saved in `uns`.
}
