% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{results.scbench}
\alias{results.scbench}
\title{Get summary of benchmark results for an analysis type}
\usage{
\method{results}{scbench}(scbench)
}
\arguments{
\item{scbench}{a processed `scbench` object}
}
\value{
a summarized table of population benchmarking results for all tested
methods, ordered by mean correlation with ground truth proportions.
}
\description{
Get summary of benchmark results for an analysis type
}
