% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaden.R
\name{scaden_scref}
\alias{scaden_scref}
\title{Compute reference matrix from an `screference` object using scaden (Single-cell assisted deconvolutional network)}
\usage{
scaden_scref(
  scref,
  cache_path = "scaden",
  bulk_mat = NULL,
  force_retrain = FALSE,
  n_cells_sim = 100,
  n_samples_sim = 1000,
  batch_size = 128,
  learning_rate = 1e-04,
  steps = 1000,
  seed = 0,
  gpu = TRUE
)
}
\arguments{
\item{scref}{an object of class `screference`}

\item{cache_path}{the path to the directory where the intermediate files
will be stored (default is "scaden")}

\item{bulk_mat}{a matrix containing the bulk gene expression data.
If NULL, a pseudobulk will be created from the reference.}

\item{force_retrain}{If TRUE, the model will be retrained even if a
trained model already exists (default is FALSE)}

\item{n_cells_sim}{the number of cells to simulate (default is 100)}

\item{n_samples_sim}{the number of samples to simulate (default is 1000)}

\item{batch_size}{the batch size used during training (default is 128)}

\item{learning_rate}{the learning rate used during training (default is 0.0001)}

\item{steps}{the number of training steps (default is 1000)}

\item{seed}{the random seed used during training (default is 0)}

\item{gpu}{if TRUE, installs Tensorflow and uses GPU acceleration (default is FALSE)}
}
\value{
The path to the directory containing the trained model
}
\description{
Compute reference matrix from an `screference` object using scaden (Single-cell assisted deconvolutional network)
}
\note{
Reference: Menden, Kevin, Mohamed Marouf, Sergio Oller et al., 2020.
“Deep Learning–Based Cell Composition Analysis from Tissue Expression Profiles.”
Science Advances 6 (30): eaba2619. https://doi.org/10.1126/sciadv.aba2619.

See also: https://github.com/theislab/AutoGeneS
}
