% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_reticulate.R
\name{seurat_to_adata}
\alias{seurat_to_adata}
\title{Convert Seurat to a adata}
\usage{
seurat_to_adata(seurat_obj, counts_assay = 1)
}
\arguments{
\item{seurat_obj}{A `Seurat` object}

\item{counts_assay}{name of assay containing counts information. Defaults to
first assay found.}
}
\value{
A converted adata object
}
\description{
This function reads a h5ad file, extracts metadata, gene and cell names, and count data.
Then, it creates a Seurat object with the extracted information.
}
\details{
Note that this version doesn't import unstructured data saved in `uns`.
}
