% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacet.R
\name{spacet_deconvolute}
\alias{spacet_deconvolute}
\title{SpaCET deconvolution using an `screference`}
\usage{
spacet_deconvolute(
  spatial_obj,
  scref,
  cancerType,
  sc_lineage_tree = NULL,
  sc_nCellEachLineage = 100,
  ncores = parallel::detectCores() - 1
)
}
\arguments{
\item{spatial_obj}{an `Seurat` object with a Spatial assay}

\item{scref}{an `screference` object containing single-cell reference data}

\item{cancerType}{A character string specifying the type of cancer.}

\item{sc_lineage_tree}{An optional data frame representing the single-cell lineage tree. Default is NULL.}

\item{sc_nCellEachLineage}{An integer specifying the number of cells for each lineage. Default is 100.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing. Default is \code{parallel::detectCores() - 1}.}
}
\description{
SpaCET deconvolution using an `screference`
}
\details{
This function utilizes the SpaCET package to perform deconvolution.
It first creates a SpaCET object with spatial transcriptomics data, and then
performs deconvolution using matched single-cell RNA-seq data.

For detailed documentation on the underlying SpaCET functions, refer to the original SpaCET package documentation:
- \code{\link{create.SpaCET.object}}
- \code{\link{SpaCET.deconvolution.matched.scRNAseq}}
}
