% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{spillover_proportions}
\alias{spillover_proportions}
\title{Simulate proportions for measuring spillover between all pairs of populations
in an `scbench` object}
\usage{
spillover_proportions(scbench, step = 0.05)
}
\arguments{
\item{scbench}{an `scbench` object}

\item{step}{a float, defines the step in the sequence for creating mixtures
and simulating spillover effects}
}
\value{
an object of class `scbench`
}
\description{
Simulate proportions for measuring spillover between all pairs of populations
in an `scbench` object
}
