% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotlight.R
\name{spotlight_deconvolute}
\alias{spotlight_deconvolute}
\title{SPOTlight deconvolution of spatial data using an `screference`}
\usage{
spotlight_deconvolute(
  spatial_obj,
  scref,
  model = "ns",
  n_top = NULL,
  min_prop = 0,
  cache_path = NULL
)
}
\arguments{
\item{spatial_obj}{an `Seurat` object with a Spatial assay}

\item{scref}{an object of `screference`}

\item{model}{character string indicating which model to use when running NMF.
Either "ns" (default) or "std".}

\item{n_top}{integer scalar specifying the number of markers to select per
group. By default NULL uses all the marker genes to initialize the model.}

\item{min_prop}{scalar in [0,1] setting the minimum contribution expected from
a cell type in x to observations in y. By default 0.}

\item{cache_path}{a path to cache the results}
}
\value{
a tibble with deconvolution fractions
}
\description{
SPOTlight deconvolution of spatial data using an `screference`
}
