% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotlight.R
\name{spotlight_scref}
\alias{spotlight_scref}
\title{Select markers for SPOTlight using an `screference`}
\usage{
spotlight_scref(
  scref,
  n_hvg = 3000,
  marker_selection_method = "mean.AUC",
  marker_selection_params = list(min = 0.8, max = Inf),
  downsample_n_cells = 100
)
}
\arguments{
\item{scref}{an `screference` object}

\item{n_hvg}{number of highly variable genes to be evaluated}

\item{marker_selection_method}{a string with the name of a metric, indicating
how to select markers. See `scran::scoreMarkers` for available metrics.}

\item{marker_selection_params}{a list of parameters to select marker genes.
`min` and `max` are currently supported. For example, for
`marker_selection_method = "mean.AUC"`, `min = 0.8` indicates markers have
minimum mean AUC of 0.8, while `max = Inf` indicates no maximum value of
mean AUC.}

\item{downsample_n_cells}{an integer, how many cells of each type to keep
in the reference}
}
\value{
reference results to be used for SPOTlight deconvolution.
}
\description{
Select markers for SPOTlight using an `screference`
}
