% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{test_markers}
\alias{test_markers}
\title{Evaluate annotation labels using Seurat markers}
\usage{
test_markers(seurat_obj, annotation, min_FC = 5, ideal_FC = 20, ncores = 4)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell gene expression data}

\item{annotation}{A character string indicating the name of the column in \code{seurat_obj@meta.data} containing the annotation of interest.}

\item{min_FC}{Minimum fold change required for a marker to pass the diagnosis. The default value is 5.}

\item{ideal_FC}{Ideal fold change required for a marker to be considered outstanding. The default value is 20.}

\item{ncores}{Number of cores to use in parallelization for the \code{FindAllMarkers} function. The default value is 4.}
}
\value{
A list of three items:
 * top_mks: data.frame with top 10 markers for each cluster, with their
 average fold change and p-value.
 * full_mks: data.frame with all markers tested, with their average fold
 change, p-value, and other metrics.
 * mk_status: named vector with the status of each cluster based on the
 expression of their top markers.
}
\description{
Identifies differentially expressed markers between clusters and evaluates their performance.
}
