% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screference.R
\name{compute_reference.screference}
\alias{compute_reference.screference}
\alias{compute_reference.hscreference}
\title{Compute reference given a deconvolution method from an `screference` object}
\usage{
\method{compute_reference}{screference}(scref, method = deconvolution_methods()[1], ...)

\method{compute_reference}{hscreference}(hscref, method = deconvolution_methods()[1], ...)
}
\arguments{
\item{scref}{an `screference` object}

\item{method}{a string, one of `deconvolution_methods()`}

\item{...}{passed to deconvolution method's reference computation function.
See the wrapper (`(method)_scref`) where `(method)` is the method name for
parameters.}

\item{hscref}{an `h-screference` object}
}
\value{
an object of class `screference`

an object of class `hscreference`
}
\description{
Compute reference given a deconvolution method from an `screference` object

Compute reference given a deconvolution method from an `h-screference` object
}
