% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconverse.R
\docType{package}
\name{deconverse-package}
\alias{deconverse}
\alias{deconverse-package}
\title{deconverse: Bulk deconvolution methods from single-cell references}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package implements wrappers for methods of bulk RNA-seq deconvolution using single-cell data to build cell population profiles. Using pseudobulk simulations, deconverse also allows for benchmarking and comparing different deconvolution methods in a context-specific manner.
}
\keyword{internal}
