% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolute_all.scbench}
\alias{deconvolute_all.scbench}
\alias{deconvolute_all.matrix}
\alias{deconvolute_all.Seurat}
\title{Deconvolute `scbench` object using all methods with default settings}
\usage{
\method{deconvolute_all}{scbench}(scbench, scref, methods = deconvolution_methods(), ...)

\method{deconvolute_all}{matrix}(bulk_data, scref, methods = deconvolution_methods(), ...)

\method{deconvolute_all}{Seurat}(spatial_obj, scref, methods = deconvolution_methods(), ...)
}
\arguments{
\item{scbench}{an `scbench` object already processed by `pseudobulks`}

\item{scref}{an `screference` object containing references for all methods that
require them}

\item{methods}{a vector of strings, which methods to use. For available
methods, consult `deconvolution_methods()`}

\item{...}{passed to `deconvolute` for CIBERSORTx credentials}

\item{bulk_data}{a count matrix of genes-by-samples}
}
\value{
an object of class `scbench`

a list with each method's deconvolution results

a list with each method's deconvolution results
}
\description{
Deconvolute `scbench` object using all methods with default settings

Deconvolute bulk RNA-seq matrix using all methods with default settings

Deconvolute Seurat spatial object using all methods with default settings
}
