% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{new_scbench}
\alias{new_scbench}
\title{Generate `scbench` object}
\usage{
new_scbench(
  ref_scrna,
  pop_bounds,
  annot_ids,
  project_name = "project",
  cache_path = "scbench_cache",
  batch_id = NULL
)
}
\arguments{
\item{ref_scrna}{a reference Seurat object}

\item{pop_bounds}{list of population bounds, see example}

\item{annot_ids}{named vector with column names of annotations per level,
e.g. `c("l1" = "annot1", "l2" = "annot2")`}

\item{project_name}{a string indicating the project name, used for caching}

\item{cache_path}{path to the directory where results will be cached}

\item{batch_id}{NULL or a string of a column name with the batch information
for simulating pseudobulks by batch}
}
\value{
an object of class scbench
}
\description{
Generate `scbench` object
}
