% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbench.R
\name{plt_spillover_scatter}
\alias{plt_spillover_scatter}
\title{Plot correlations between deconvolution results and spillover mixtures
between each pair of populations}
\usage{
plt_spillover_scatter(scbench, method, level = NULL)
}
\arguments{
\item{scbench}{an `scbench` object that has been evaluated by
`deconvolute`}

\item{method}{a string, one of `deconvolution_methods()`}

\item{level}{an annotation level to show. Finest-grained annotation (highest level)
by default.}
}
\value{
a ggplot object
}
\description{
Plot correlations between deconvolution results and spillover mixtures
between each pair of populations
}
